Move string truncation to its own package
