+++ upup/pkg/fi/cloudup/awsup/BUILD.bazel
@@ -27,6 +27,7 @@ go_library(
         "//pkg/featureflag:go_default_library",
         "//pkg/nodeidentity/aws:go_default_library",
         "//pkg/resources/spotinst:go_default_library",
+        "//pkg/truncate:go_default_library",
         "//protokube/pkg/etcd:go_default_library",
         "//upup/pkg/fi:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
