+++ upup/pkg/fi/cloudup/awsup/aws_utils.go
@@ -17,9 +17,7 @@ limitations under the License.
 package awsup
 
 import (
-	"encoding/base32"
 	"fmt"
-	"hash/fnv"
 	"os"
 	"strings"
 	"sync"
@@ -35,6 +33,7 @@ import (
 	elbv2 "github.com/aws/aws-sdk-go/service/elbv2"
 	"k8s.io/klog/v2"
 	"k8s.io/kops/pkg/apis/kops"
+	"k8s.io/kops/pkg/truncate"
 )
 
 // allRegions is the list of all regions; tests will set the values
@@ -223,7 +222,7 @@ func ELBv2Tags(tags map[string]string) []*elbv2.Tag {
 
 // GetClusterName40 will attempt to calculate a meaningful cluster name with a max length of 40
 func GetClusterName40(cluster string) string {
-	return TruncateString(cluster, TruncateStringOptions{
+	return truncate.TruncateString(cluster, truncate.TruncateStringOptions{
 		MaxLength: 40,
 	})
 }
@@ -234,60 +233,12 @@ func GetResourceName32(cluster string, prefix string) string {
 	s := prefix + "-" + strings.Replace(cluster, ".", "-", -1)
 
 	// We always compute the hash and add it, lest we trick users into assuming that we never do this
-	opt := TruncateStringOptions{
+	opt := truncate.TruncateStringOptions{
 		MaxLength:     32,
 		AlwaysAddHash: true,
 		HashLength:    6,
 	}
-	return TruncateString(s, opt)
-}
-
-// TruncateStringOptions contains parameters for how we truncate strings
-type TruncateStringOptions struct {
-	// AlwaysAddHash will always cause the hash to be appended.
-	// Useful to stop users assuming that the name will never be truncated.
-	AlwaysAddHash bool
-
-	// MaxLength controls the maximum length of the string.
-	MaxLength int
-
-	// HashLength controls the length of the hash to be appended.
-	HashLength int
-}
-
-// TruncateString will attempt to truncate a string to a max, adding a prefix to avoid collisions.
-// Will never return a string longer than maxLength chars
-func TruncateString(s string, opt TruncateStringOptions) string {
-	if opt.MaxLength == 0 {
-		klog.Fatalf("MaxLength must be set")
-	}
-
-	if !opt.AlwaysAddHash && len(s) <= opt.MaxLength {
-		return s
-	}
-
-	if opt.HashLength == 0 {
-		opt.HashLength = 6
-	}
-
-	// We always compute the hash and add it, lest we trick users into assuming that we never do this
-	h := fnv.New32a()
-	if _, err := h.Write([]byte(s)); err != nil {
-		klog.Fatalf("error hashing values: %v", err)
-	}
-	hashString := base32.HexEncoding.EncodeToString(h.Sum(nil))
-	hashString = strings.ToLower(hashString)
-	if len(hashString) > opt.HashLength {
-		hashString = hashString[:opt.HashLength]
-	}
-
-	maxBaseLength := opt.MaxLength - len(hashString) - 1
-	if len(s) > maxBaseLength {
-		s = s[:maxBaseLength]
-	}
-	s = s + "-" + hashString
-
-	return s
+	return truncate.TruncateString(s, opt)
 }
 
 // GetTargetGroupNameFromARN will attempt to parse a target group ARN and return its name
